requirements.txt
django
tornado
python-dbusx